Version Bump and adding documentation
The version bump is because, when adding javadocs, I took the oportunity
to change the names of a bunch of methods, so strictly, this is an
incompatable API change.

On the other hand, this is an application, not a library, so I can do
what I like with my internal APIs.